Move loadWithTs to CellLoader
There is some log-based logic here (isTraceEnabled); since we're in the
same package, I'm assuming that the same rules will apply. Is
assumption this valid in practice?
